+++ lib/Braintree/Test/TransactionAmounts.php
@@ -10,7 +10,8 @@ namespace Braintree\Test;
  */
 class TransactionAmounts
 {
-    public static $authorize   = '1000.00';
-    public static $decline     = '2000.00';
-    public static $hardDecline = '2015.00';
+    public static $authorize           = '1000.00';
+    public static $partiallyAuthorized = '1004.00';
+    public static $decline             = '2000.00';
+    public static $hardDecline         = '2015.00';
 }